IPCThreadState: Add a public method to probe if a binder call is being served.
This can be useful when common code can be executed by both hwbinder and
binder threads, and, they want to access (hardware::)IPCThreadState
specific information. Eg: cameraserver.

This commit adds a vndk private library 'libbinderthreadstate' which
serves as common ground for both libbinder and libhwbinder to inform
them about the nature of the IPC call being served (if at all) on the current
thread.

Bug: 110364143

Test: Boot device, use GCA

Change-Id: Ifefb273c9a1f545417a82e9511a45bda00f701aa
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>