libbinder: +2 bytes in BBinder from stability rep
sizeof(BBinder) may not be changed unless 1,000s of people in many
different companies fundamentally change the way they work. So, with
precious few bits to spare, we make room by changing the way that
binder stability reserves space on the wire. Now, it uses the least
significant 16-bits of the 32-bits which is reserved on the wire.

The sideeffect of this straightforward implementation is that the wire
protocol is slightly changed. This is an intentional change in order to
exercise its instability, perhaps as an early warning.

Bug: 166282674
Test: boot, binderLibTest

Change-Id: I654fcd2cc9d20cbac557d1a176a5095c491d88cf