binder: Increase RPC backlog to 50.
Right now, we create all threads at once, making accept4 slow.
To avoid hanging the client, the backlog is increased to a large number.
Once we create threads dynamically & lazily, the backlog can be reduced
to 1.

In an experiment, `time aservice list` reduces from 4 minutes to 30
seconds, a 8x boost.

Bug: 189955605
Bug: 190868305
Test: `time aservice list` is significantly faster

Change-Id: I4fa4d62149891ab62cbaf9f688bf798e4faf0189